Exported Types

The Wind Farm Simulation struct

FLORIDyn.WindFarm — Type

WindFarm

A mutable struct representing a wind farm. Fields can be specified using keyword arguments.

Fields

nT::Int64: Number of turbines

nOP::Int64: Number of operating points

States_WF::Matrix{Float64}: States of the wind farm

States_OP::Matrix{Float64}: States of the operating points

States_T::Matrix{Float64}: States of the turbines

posBase::Matrix{Float64}: Base positions of the turbines

posNac::Matrix{Float64}: Positions of the nacelles

D::Vector{Float64}: Diameters of the turbines

StartI::Matrix{Int}: Start indices for each turbine

intOPs::Vector{Matrix{Float64}}: Interpolated operating points

Weight::Vector{Vector{Float64}}: Weights for the operating points

dep::Vector{Vector{Int}}: Dependencies between turbines

red_arr::Matrix{Float64}: Reduced array for each turbine

Names_T::Vector{String}: Names of the states of the turbines

Names_WF::Vector{String}: Names of the states of the wind farm

Names_OP::Vector{String}: Names of coordinates the operating points

Markers defining the wind velocity model

An instance of these structs needs to be passed to the functions that calculate the wind velocity. They are all subtypes of VelModel

FLORIDyn.Velocity_Constant — Type

Velocity_Constant <: VelModel

A velocity model representing a constant wind velocity field. This struct is used as a type marker to indicate that the wind velocity does not vary in space or time.

See also:

VelModel: Abstract supertype for velocity models.

FLORIDyn.Velocity_Constant_wErrorCov — Type

Velocity_Constant_wErrorCov

A velocity model representing a constant wind field with associated error covariance. This struct is a subtype of VelModel and is used to model wind velocity with an constant value and an error covariance for uncertainty quantification.

See also:

VelModel: Abstract supertype for velocity models.

FLORIDyn.Velocity_EnKF_InterpTurbine — Type

Velocity_EnKF_InterpTurbine

A velocity model type representing an interpolated turbine velocity field using the Ensemble Kalman Filter (EnKF) approach.

Description

This struct is used within the wind field modeling framework to represent the velocity at a turbine location, where the velocity is estimated or interpolated using EnKF-based techniques.

See also:

VelModel: Abstract supertype for velocity models.

Markers defining the wind direction model

An instance of these structs needs to be passed to the functions that calculate the wind direction. They are all subtypes of DirModel

FLORIDyn.Direction_Constant — Type

Direction_Constant <: DirModel

A marker struct used to represent a constant wind direction.

Example:

dir_mode = Direction_constant()
phi = getWindDirT(dir_mode, 270, [1,2,3], nothing)

FLORIDyn.Direction_Constant_wErrorCov — Type

Direction_Constant_wErrorCov <: DirModel

A marker struct used to indicate a wind direction that is constant with associated error covariance.

FLORIDyn.Direction_EnKF_InterpTurbine — Type

Direction_EnKF_InterpTurbine <: DirModel

A marker struct used to indicate the use of direction-aware Ensemble Kalman Filter (EnKF) interpolation for turbine modeling.

FLORIDyn.Direction_Interpolation — Type

Direction_Interpolation <: DirModel

A marker struct used to represent direction interpolation functionality within the FLORIDyn framework.

FLORIDyn.Direction_Interpolation_wErrorCov — Type

Direction_Interpolation_wErrorCov <: DirModel

A marker struct representing a direction interpolation method with associated error covariance.

FLORIDyn.Direction_InterpTurbine — Type

Direction_InterpTurbine <: DirModel

A marker struct used to indicate direction interpolation for turbines.

FLORIDyn.Direction_InterpTurbine_wErrorCov — Type

Direction_InterpTurbine_wErrorCov <: DirModel

A marker struct used to indicate the use of direction interpolation for turbines with associated error covariance.

FLORIDyn.Direction_RW_with_Mean — Type

Direction_RW_with_Mean <: DirModel

A marker struct used to indicate the use of a random walk direction model with a mean component.

Markers for defining the wind shear

An instance of these structs needs to be passed to the functions that calculate the wind shear.

FLORIDyn.Shear_Interpolation — Type

Shear_Interpolation

A marker struct used to represent the linear interpolation for wind shear profiles.

See also

Shear_PowerLaw

FLORIDyn.Shear_PowerLaw — Type

Shear_PowerLaw

A marker struct representing the logarithmic law for modeling wind shear profiles.

See also

Shear_Interpolation

Markers for defining the wind turbulence

An instance of these structs needs to be passed to the functions that calculate the wind turbulence.

FLORIDyn.TI_Constant — Type

TI_Constant

A marker struct representing a constant turbulence intensity.

FLORIDyn.TI_EnKF_InterpTurbine — Type

TI_EnKF_InterpTurbine

A marker struct representing the Turbulence Intensity (TI) Ensemble Kalman Filter (EnKF) interpolation model.

FLORIDyn.TI_Interpolation — Type

TI_Interpolation

A marker struct representing the interpolation method for modeling the turbulence.

FLORIDyn.TI_InterpTurbine — Type

TI_InterpTurbine

A marker struct representing an interpolated turbine model for turbulence intensity calculations.

Types for storing wind field params

FLORIDyn.WindDirType — Type

 WindDirType

Fields

Data::Float64: wind direction value

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindDirMatrix — Type

struct WindDirMatrix

Fields

Data::Matrix{Float64}: Columns [time, phi] or [time, phiT0, phiT1, ... phi_Tn]

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindDirTriple — Type

WindDirTriple

A structure representing a wind direction triple.

Fields

Init::Vector{Float64}: Mean direction (vector or scalar)

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

MeanPull::Float64: Scalar mean reversion factor

FLORIDyn.WindVelType — Type

 WindVelType

Fields

Data::Float64: wind speed

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindVelMatrix — Type

struct WindVelMatrix

Fields

Data::Matrix{Float64}: Nx2 matrix: column 1 = time, column 2 = wind speed

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)