Exported Types

Markers defining the wind direction model

An instance of these structs needs to be passed to the functions that calculate the wind direction. They are all subtypes of DirModel

FLORIDyn.Direction_Constant — Type

Direction_Constant <: DirModel

A marker struct used to represent a constant wind direction.

Example:

dir_mode = Direction_constant()
phi = getWindDirT(dir_mode, 270, [1,2,3], nothing)

FLORIDyn.Direction_Constant_wErrorCov — Type

Direction_Constant_wErrorCov <: DirModel

A marker struct used to indicate a wind direction that is constant with associated error covariance.

FLORIDyn.Direction_EnKF_InterpTurbine — Type

Direction_EnKF_InterpTurbine <: DirModel

A marker struct used to indicate the use of direction-aware Ensemble Kalman Filter (EnKF) interpolation for turbine modeling.

FLORIDyn.Direction_Interpolation — Type

Direction_Interpolation <: DirModel

A marker struct used to represent direction interpolation functionality within the FLORIDyn framework.

FLORIDyn.Direction_Interpolation_wErrorCov — Type

Direction_Interpolation_wErrorCov <: DirModel

A marker struct representing a direction interpolation method with associated error covariance.

FLORIDyn.Direction_InterpTurbine — Type

Direction_InterpTurbine <: DirModel

A marker struct used to indicate direction interpolation for turbines.

FLORIDyn.Direction_InterpTurbine_wErrorCov — Type

Direction_InterpTurbine_wErrorCov <: DirModel

A marker struct used to indicate the use of direction interpolation for turbines with associated error covariance.

FLORIDyn.Direction_RW_with_Mean — Type

Direction_RW_with_Mean <: DirModel

A marker struct used to indicate the use of a random walk direction model with a mean component.

Markers for defining the wind shear

An instance of these structs needs to be passed to the functions that calculate the wind shear.

FLORIDyn.Shear_Interpolation — Type

Shear_Interpolation

A marker struct used to represent the linear interpolation for wind shear profiles.

See also

Shear_PowerLaw

FLORIDyn.Shear_PowerLaw — Type

Shear_PowerLaw

A marker struct representing the logarithmic law for modeling wind shear profiles.

See also

Shear_Interpolation

Markers for defining the wind turbulence

An instance of these structs needs to be passed to the functions that calculate the wind turbulence.

FLORIDyn.TI_Constant — Type

TI_Constant

A marker struct representing a constant turbulence intensity.

FLORIDyn.TI_EnKF_InterpTurbine — Type

TI_EnKF_InterpTurbine

A marker struct representing the Turbulence Intensity (TI) Ensemble Kalman Filter (EnKF) interpolation model.

FLORIDyn.TI_Interpolation — Type

TI_Interpolation

A marker struct representing the interpolation method for modeling the turbulence.

FLORIDyn.TI_InterpTurbine — Type

TI_InterpTurbine

A marker struct representing an interpolated turbine model for turbulence intensity calculations.

Types for storing wind field params

FLORIDyn.WindDirType — Type

 WindDirType

Fields

Data::Float64: wind direction value

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindDirMatrix — Type

struct WindDirMatrix

Fields

Data::Matrix{Float64}: Nx2 matrix: column 1 = time, column 2 = phi

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindDirTriple — Type

WindDirTriple

A structure representing a wind direction triple.

Fields

Init::Vector{Float64}: Mean direction (vector or scalar)

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

MeanPull::Float64: Scalar mean reversion factor

FLORIDyn.WindVelType — Type

 WindVelType

Fields

Data::Float64: wind speed

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)

FLORIDyn.WindVelMatrix — Type

struct WindVelMatrix

Fields

Data::Matrix{Float64}: Nx2 matrix: column 1 = time, column 2 = wind speed

CholSig::Matrix{Float64}: Cholesky factor of covariance matrix (nT x nT)